main

+++ code/Encoder.h
@@ -15,7 +15,7 @@ private:
     long* CNT;//编码器对应GPIO通道
 
     int carry_flag = 0;
-    uint32_t prev_cnt = 0;
+    long prev_cnt = 0;
 
     bool init_flag = false;//初始化是否完成，初始化处理和一般处理不同
